requirements.txt
torch
transformers
Pillow
requests
pytest
pytest-html
